Add files via upload
W.I.P English translation. Mos, make sure to go through and correct any mistranslations. Not 100% accurate. All lines should be the same as the original Chinese MD file. Mostly adding this as a test for when you implement the version change, so I can see how this will be formatted on the blog, and fix it accordingly. Will update as I go.